Rename _forward/prism.py to _forward/prism_gravity.py
