+++ harmonica/tests/test_prism.py
@@ -32,7 +32,7 @@ except ImportError:
     ProgressBar = None
 
 from .. import bouguer_correction
-from .._forward.prism import _check_prisms, _discard_null_prisms, prism_gravity
+from .._forward.prism_gravity import _check_prisms, _discard_null_prisms, prism_gravity
 from .utils import run_only_with_numba
